+++ ttn/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## [2.1.2] - 2025-11-27
+
+### Add
+
+- Added support for enabling/disabling individual applications in the backend
+- Improved Inspector messages
+- Auto completion for application edition in the plugin configuration page
+
+### Fix
+
+- Fixed multiple application selection bug in the plugin configuration page
+- Fixed edition of existing plugin applications
+
+## [2.1.1] - 2025-11-13
+
+### Add
+
+- Changed visual style of plugin sections in order to improve readability
+- Significantly improved error handling and logging for uplink parsing issues
+
 ## [2.1.0] - 2025-11-12
 
 ### Add